feat: add 'Admin mode' to view tomorrow's game
Lets administrator enter a password to get access to tomorrow's game
to easier fix bugs. Obviously, I'm aware that having that password
visible in the github repo is not exactly secure, but since the only
thing "admin mode" contains here is a way to get one game ahead of time,
it really doesn't matter too much if anyone decides to abuse it. An
eventual expansion of admin mode would obviously include increased
security.